Refactor the report module into a package
