+++ src/morphoclass/console/stats_plotter.py
@@ -25,7 +25,8 @@ from matplotlib.figure import Figure
 from morphio import Morphology
 from tmd.Topology.methods import get_persistence_diagram
 
-from morphoclass import report
+import morphoclass as mc
+import morphoclass.report.plumbing
 from morphoclass.vis import plot_counts_per_layer
 from morphoclass.vis import plot_counts_per_subclass
 from morphoclass.vis import plot_histogram
@@ -231,6 +232,6 @@ class DatasetStatsPlotter:
     def save_report(self, report_title: str = "Dataset Plots") -> None:
         """Save the HTML report to disk."""
         self.template_vars["report_title"] = report_title
-        template = report.load_template("plot-dataset-stats")
+        template = mc.report.plumbing.load_template("plot-dataset-stats")
         self.report_dir.mkdir(exist_ok=True, parents=True)
-        report.render(template, self.template_vars, self.report_path)
+        mc.report.plumbing.render(template, self.template_vars, self.report_path)
