+++ src/morphoclass/report/__init__.py
@@ -1,0 +1,2 @@
+"""Tools for creating jinja-based HTML reports."""
+from __future__ import annotations
